data preparation/requirements.txt
PyPDF2==3.0.1
pandas==2.1.4
openpyxl==3.1.2
nltk==3.8.1
beautifulsoup4==4.12.2
spacy==3.7.2
emoji==2.8.0
regex==2023.10.3
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.13
faiss-cpu==1.7.4
torch==2.1.2
scikit-learn==1.3.2
numpy==1.26.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1.tar.gz 